sharding/pyproject.toml
[tool.poetry]
name = "sharding"
version = "0.1.0"
description = ""
authors = ["Ihor <cguilliman@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
Faker = "^22.2.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
